+++ packages/ai/core/generate-text/run-tools-transformation.ts
@@ -47,6 +47,10 @@ export type SingleRequestTextStreamPart<TOOLS extends ToolSet> =
   | {
       type: 'error';
       error: unknown;
+    }
+  | {
+      type: 'raw';
+      rawValue: unknown;
     };
 
 export function runToolsTransformation<TOOLS extends ToolSet>({
@@ -128,6 +132,12 @@ export function runToolsTransformation<TOOLS extends ToolSet>({
         case 'source':
         case 'response-metadata':
         case 'error': {
+          controller.enqueue(chunk);
+          break;
+        }
+
+        // forward raw chunks to be part of the fullStream
+        case 'raw': {
           controller.enqueue(chunk);
           break;
         }