+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @ai-sdk/google-vertex
 
+## 3.0.0-alpha.12
+
+### Patch Changes
+
+- e2aceaf: feat: add raw chunk support
+- Updated dependencies [e2aceaf]
+  - @ai-sdk/anthropic@2.0.0-alpha.12
+  - @ai-sdk/google@2.0.0-alpha.12
+  - @ai-sdk/provider@2.0.0-alpha.12
+  - @ai-sdk/provider-utils@3.0.0-alpha.12
+
 ## 3.0.0-alpha.11
 
 ### Patch Changes