+++ packages/rsc/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @ai-sdk/rsc
 
+## 1.0.0-alpha.12
+
+### Patch Changes
+
+- Updated dependencies [da1e6f0]
+- Updated dependencies [e2aceaf]
+- Updated dependencies [4892798]
+  - ai@5.0.0-alpha.12
+  - @ai-sdk/provider@2.0.0-alpha.12
+  - @ai-sdk/provider-utils@3.0.0-alpha.12
+
 ## 1.0.0-alpha.11
 
 ### Patch Changes